Revert "Merge branch 'main' into gn-field-statistic-node"
This reverts commit 2781f02c7fe7d05b0f88aff2c6619281a3a4b051.

+++ source/blender/io/usd/intern/usd_hash_types.hh
@@ -28,7 +28,7 @@ template<> struct DefaultHash<pxr::TfToken> {
 template<> struct DefaultHash<pxr::SdfPath> {
   uint64_t operator()(const pxr::SdfPath &value) const
   {
-    return value.GetHash();
+    return (uint64_t)value.GetHash();
   }
 };
 }  // namespace blender
